% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uggs.R
\name{uggs}
\alias{uggs}
\title{Nonparametric bias-corrected and accelerated bootstrap confidence
  limits}
\usage{
uggs(df, B, est, ..., jcount = nrow(df), jreps = 5, iereps = 2,
  J = 10, alpha = c(0.025, 0.05, 0.1), progress = TRUE,
  num.workers = 4)
}
\description{
This routine computes nonparametric confidence limits for
  bootstrap estimates.
}
\details{
Bootstrap confidence limits correct the standard method of
confidence intervals in three ways

- the bootstrap cdf, \eqn{G} - the bias-correction number \eqn{<<latex>>}{<<plain>>}{eqn{G}} - the
acceleration number ${a}$ that measures the rate of change in
\eqn{\sigma_{t_0}} as the data changes.
}
